core: simplify guardian review session lifecycle
